Fix generated man page and MacOS librsync (#808)
FIX: man pages are pre-generated from asciidoc to avoid one dependency at package time and avoid trying to package librsync.dll under MacOS, closes #807
CHG: the man pages generated from asciidoc are available under dist/ instead of build/ (that should be only relevant to packagers)

+++ setup.py
@@ -10,6 +10,7 @@ import time
 # we need all this to extend the distutils/setuptools commands
 from setuptools import setup, Extension, Command
 import setuptools.command.build_py
+import setuptools.command.sdist
 from distutils.debug import DEBUG
 import distutils.command.clean
 from distutils import log
@@ -42,24 +43,25 @@ if os.name == "posix" or os.name == "nt":
         if LFLAGS or LIBS:
             lflags_arg = LFLAGS + LIBS
 
-        if LIBRSYNC_DIR and len(sys.argv) > 1:  # should only be under Windows
+        if LIBRSYNC_DIR and len(sys.argv) > 1:
             incdir_list = [os.path.join(LIBRSYNC_DIR, "include")]
             libdir_list = [os.path.join(LIBRSYNC_DIR, "lib")]
-            rsyncdll_src = os.path.join(LIBRSYNC_DIR, "bin", "rsync.dll")
-            rsyncdll_dst = os.path.join("src", "rdiff_backup", "rsync.dll")
-            # rather ugly workaround, but it should be good enough
-            if "clean" in sys.argv:
-                if os.path.exists(rsyncdll_dst) and "--all" in sys.argv:
-                    print(f"removing {rsyncdll_dst}")
-                    if "--dry-run" not in sys.argv:
-                        os.remove(rsyncdll_dst)
-            elif ("--version" not in sys.argv and "-V" not in sys.argv
-                  and "--help" not in sys.argv):
-                if (not os.path.exists(rsyncdll_dst)
-                        or not filecmp.cmp(rsyncdll_src, rsyncdll_dst)):
-                    print(f"copying {rsyncdll_src} -> {rsyncdll_dst}")
-                    if "--dry-run" not in sys.argv:
-                        shutil.copyfile(rsyncdll_src, rsyncdll_dst)
+            if os.name == "nt":
+                rsyncdll_src = os.path.join(LIBRSYNC_DIR, "bin", "rsync.dll")
+                rsyncdll_dst = os.path.join("src", "rdiff_backup", "rsync.dll")
+                # rather ugly workaround, but it should be good enough
+                if "clean" in sys.argv:
+                    if os.path.exists(rsyncdll_dst) and "--all" in sys.argv:
+                        print(f"removing {rsyncdll_dst}")
+                        if "--dry-run" not in sys.argv:
+                            os.remove(rsyncdll_dst)
+                elif ("--version" not in sys.argv and "-V" not in sys.argv
+                      and "--help" not in sys.argv):
+                    if (not os.path.exists(rsyncdll_dst)
+                            or not filecmp.cmp(rsyncdll_src, rsyncdll_dst)):
+                        print(f"copying {rsyncdll_src} -> {rsyncdll_dst}")
+                        if "--dry-run" not in sys.argv:
+                            shutil.copyfile(rsyncdll_src, rsyncdll_dst)
         if "-lrsync" in LIBS:
             libname = []
 
@@ -71,7 +73,7 @@ if os.name == "nt":
 # --- extend the build command to execute a command ---
 
 
-class build_exec(Command):
+class pre_build_exec(Command):
     description = 'build template files executing a shell command'
     user_options = [
         # The format is (long option, short option, description).
@@ -120,7 +122,7 @@ class build_exec(Command):
 # --- extend the build command to do templating of files ---
 
 
-class build_templates(Command):
+class pre_build_templates(Command):
     description = 'build template files replacing {{ }} placeholders'
     user_options = [
         # The format is (long option, short option, description).
@@ -171,12 +173,21 @@ class build_templates(Command):
 
 
 class build_py(setuptools.command.build_py.build_py):
-    """Inject our build sub-command in the build step"""
+    """Inject our build sub-command in the build_py step"""
 
     def run(self):
-        self.run_command('build_exec')
-        self.run_command('build_templates')
+        self.run_command('pre_build_exec')
+        self.run_command('pre_build_templates')
         setuptools.command.build_py.build_py.run(self)
+
+
+class sdist(setuptools.command.sdist.sdist):
+    """Inject our build sub-command in the sdist step"""
+
+    def run(self):
+        self.run_command('pre_build_exec')
+        self.run_command('pre_build_templates')
+        setuptools.command.sdist.sdist.run(self)
 
 
 # --- extend the clean command to remove templated and exec files ---
@@ -191,10 +202,10 @@ class clean(distutils.command.clean.clean):
 
     def finalize_options(self):
         """Post-process options."""
-        # take over the option from our build_templates command
-        self.set_undefined_options('build_templates',
+        # take over the option from our pre_build_templates command
+        self.set_undefined_options('pre_build_templates',
                                    ('template_files', 'template_files'))
-        self.set_undefined_options('build_exec',
+        self.set_undefined_options('pre_build_exec',
                                    ('commands', 'commands'))
         super().finalize_options()
 
@@ -283,10 +294,10 @@ setup(
         ),
     ],
     data_files=[
-        ("share/man/man1", ["build/rdiff-backup.1",
-                            "build/rdiff-backup-old.1",
-                            "build/rdiff-backup-delete.1",
-                            "build/rdiff-backup-statistics.1"]),
+        ("share/man/man1", ["dist/rdiff-backup.1",
+                            "dist/rdiff-backup-old.1",
+                            "dist/rdiff-backup-delete.1",
+                            "dist/rdiff-backup-statistics.1"]),
         (
             "share/doc/rdiff-backup", [
                 "CHANGELOG.adoc",
@@ -306,29 +317,30 @@ setup(
     # options is a hash of hash with command -> option -> value
     # the value happens here to be a list of file couples/tuples
     options={
-        'build_templates': {'template_files': [
+        'pre_build_templates': {'template_files': [
             ("tools/rdiff-backup.spec.template", "build/rdiff-backup.spec"),
             ("tools/rdiff-backup.spec.template-fedora", "build/rdiff-backup.fedora.spec"),
-            ("docs/rdiff-backup-old.1", "build/rdiff-backup-old.1"),
+            ("docs/rdiff-backup-old.1.template", "dist/rdiff-backup-old.1"),
         ]},
-        "build_exec": {"commands": [
+        "pre_build_exec": {"commands": [
             ("asciidoctor -b manpage -a revdate=\"{date}\" "
              "-a revnumber=\"{ver}\" -o {outfile} {infile}",
-             ("docs", "rdiff-backup.1.adoc"), ("build", "rdiff-backup.1")),
+             ("docs", "rdiff-backup.1.adoc"), ("dist", "rdiff-backup.1")),
             ("asciidoctor -b manpage -a revdate=\"{date}\" "
              "-a revnumber=\"{ver}\" -o {outfile} {infile}",
              ("docs", "rdiff-backup-statistics.1.adoc"),
-             ("build", "rdiff-backup-statistics.1")),
+             ("dist", "rdiff-backup-statistics.1")),
             ("asciidoctor -b manpage -a revdate=\"{date}\" "
              "-a revnumber=\"{ver}\" -o {outfile} {infile}",
              ("docs", "rdiff-backup-delete.1.adoc"),
-             ("build", "rdiff-backup-delete.1")),
+             ("dist", "rdiff-backup-delete.1")),
         ]},
     },
     cmdclass={
-        'build_exec': build_exec,
-        'build_templates': build_templates,
+        'pre_build_exec': pre_build_exec,
+        'pre_build_templates': pre_build_templates,
         'build_py': build_py,
+        'sdist': sdist,
         'clean': clean,
     },
     install_requires=[
